allowing valueFrom for using outputs from a different component
